+++ libdiskonaut/src/model/files/mod.rs
@@ -1,8 +1,10 @@
+mod contents;
 mod file_or_folder;
 mod file_tree;
 mod hard_links;
 mod hash;
 
+pub use contents::Contents;
 pub use file_or_folder::*;
 pub use file_tree::*;
 pub use hard_links::*;